perf(webpack-cli): defer interpret import and inline Levenshtein helper
Profiling `webpack build` showed CLI-owned code is a tiny fraction of
startup; the cost is module loading. Two reductions there:

- Default-config discovery now imports `interpret` only when no
  common-extension config (.js/.mjs/.cjs/.ts/.cts/.mts) is found, so the
  common `webpack.config.js` build never loads it.
- The Levenshtein "did you mean" helper is inlined into webpack-cli as a
  private method, removing the separate module/import. The 256 KB buffer
  stays lazily allocated (error paths only).
